use common base class
