+++ qiskit/transpiler/passes/__init__.py
@@ -82,7 +82,9 @@ Calibration
 .. autosummary::
    :toctree: ../stubs/
 
-    PulseGates
+   PulseGates
+   RZXCalibrationBuilder
+   RZXCalibrationBuilderNoEcho
 
 Scheduling
 =============
@@ -96,8 +98,6 @@ Scheduling
    DynamicalDecoupling
    AlignMeasures
    ValidatePulseGates
-   RZXCalibrationBuilder
-   RZXCalibrationBuilderNoEcho
 
 Circuit Analysis
 ================
@@ -197,13 +197,13 @@ from .synthesis import UnitarySynthesis
 
 # calibration
 from .calibration import PulseGates
+from .calibration import RZXCalibrationBuilder
+from .calibration import RZXCalibrationBuilderNoEcho
 
 # circuit scheduling
 from .scheduling import TimeUnitConversion
 from .scheduling import ALAPSchedule
 from .scheduling import ASAPSchedule
-from .scheduling import RZXCalibrationBuilder
-from .scheduling import RZXCalibrationBuilderNoEcho
 from .scheduling import DynamicalDecoupling
 from .scheduling import AlignMeasures
 from .scheduling import ValidatePulseGates
