+++ qiskit/transpiler/passes/calibration/__init__.py
@@ -12,4 +12,4 @@
 
 """Module containing transpiler calibration passes."""
 
-from .pulse_gates import PulseGates
+from .creators import RZXCalibrationBuilder, RZXCalibrationBuilderNoEcho, PulseGates
