+++ qiskit/transpiler/passes/scheduling/__init__.py
@@ -15,11 +15,5 @@
 from .alap import ALAPSchedule
 from .asap import ASAPSchedule
 from .time_unit_conversion import TimeUnitConversion
-from .calibration_creators import (
-    CalibrationCreator,
-    RZXCalibrationBuilder,
-    RZXCalibrationBuilderNoEcho,
-)
 from .dynamical_decoupling import DynamicalDecoupling
-from .rzx_templates import rzx_templates
 from .instruction_alignment import AlignMeasures, ValidatePulseGates
